refactor(voicecli): create api/ sub-package (T2 of #83)
5 modules moved under voicecli.api (api.py, markdown.py, _markdown_directives.py,
translate.py, engine_caps.py). voicecli/__init__.py + voicecli/api/__init__.py
re-exports preserve public API. No behavior change.

Note: plan referenced api_chunked.py and markdown_types.py — both already
consolidated upstream; nothing to move. Updated file_exemptions.txt for
api.py → api/api.py path move.
