+++ src/voicecli/engines/qwen_fast.py
@@ -20,7 +20,7 @@ from voicecli.models import (
 )
 
 if TYPE_CHECKING:
-    from voicecli.markdown import Segment
+    from voicecli.api.markdown import Segment
 
 
 class QwenFastEngine(QwenEngine):
